SimpleLink LPF2 8.32.00.07 and LPF3 9.20.00.81 coresdk examples update This commit updates coresdk examples LPF2 and LPF3 SDK versions

+++ examples/rtos/LP_EM_CC2340R53/drivers/uart2callback/uart2callback.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020, Texas Instruments Incorporated
+ * Copyright (c) 2020-2025, Texas Instruments Incorporated
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -74,7 +74,7 @@ void *mainThread(void *arg0)
     UART2_Handle uart;
     UART2_Params uartParams;
     int32_t semStatus;
-    uint32_t status = UART2_STATUS_SUCCESS;
+    int_fast16_t status = UART2_STATUS_SUCCESS;
 
     /* Call driver init functions */
     GPIO_init();
